activateApp: switch to non-method fn bind, w/ AdminWebsocket impl
we bind directly to the method in question, relying on the
first-argument being `this`. in our wrapper fn we pass in the inner,
"wrapped", websocket js value (corresponding to what holochain-client-js
calls `AdminWebsocket`).

I think this approach is better. it seems to work but I will wire up a
test next.