introduce `ZomeCallable` trait
we will use this to bind, in a type-safe way, the inputs and outputs of
zome calls. we will be able to perform conversions to/from Javascript in
a way which *should* work consistently (provided `holochain-client-js`
behaves properly).

we will export a macro which will generate impls for this trait
corresponding to `hdk_extern` functions from zomes. frontend code will
be able to import & use those impls in order to make type safe[1] calls
to the backend.

[1]: type safety here is only partial - there is the very real
possibility of runtime exceptions b/c we are interfacing w/ Javascript.
this seems like the best we can get at this stage, since
`holochain-websocket` is not prepared to target WASM. the type safety I
am describing is for the frontend author - any zome call they make will
be checked against the signature of the literal hdk_extern "backend"
function. modulo any stupid JS conversion errors, that should mean that
things will convert into payloads, be sent off as
`AppWebsocket.callZome`s, and then be converted back into the correct
return type. hopefully this will work 🤞