work around idiotic wasm-bindgen issue
not sure whose fault it is, but I find it rather annoying.

it seems that hApp Rust/wasm code is unable to depend on crates which
themselves depend of `wasm-bindgen`. this means that I cannot factor
bits of code which depend on `wasm-bindgen` into common crates. that was
about 80% of my goal with `paperz_core`, and now seems to be totally
moot.

oh well.

here, I created new `SerializeToJsObj_` & DeserializeFromJsObj_` traits
which are copies of the non-underscore-suffixed versions, in `paperz_ui`.
I and added impls for `Paper` so that I can convert those in and out of
JsValue. my wrapper types shim to the original versions of the traits,
so that we get some amount of nice reuse.

however, it required awkwardly defining my own `Pair` type, which will
only work for mixed `_`-prefixed and ``-prefixed Ser/De pairs of a
certain order. we would need to enumerate all possible combinations of
such pairs, which is absolutely tedious and stupid and I hate it.

however, it seems to work, and maybe that's alright for now.

I would, however, like to track down exactly wtf Holochain is doing
which causes these issues...